[api/triggers] Move dispatch tests to core
Keep subscriber coverage focused on envelope adaptation, and move fan-out, idempotency, and history behavior tests to the core dispatcher boundary.

Refs ENG-547
